Add support for redirecting ports with slirp
We can now redirect ports on the host to the guest VM using slirp.

To enable, specify HOST_IP and SLIRP_PORTS. For example, to enable access
to a web service running in the guest VM:

    HOST_IP=<eth0 IP> SLIRP_PORTS="80 443" ./run <some web service>

Specifying the host IP is required because slirp is bad at detecting it and
usually tries to bind to 127.0.0.1.